+++ unicode-data-names/test/ICU/NamesSpec.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE BlockArguments #-}
+{-# LANGUAGE CPP, BlockArguments #-}
 
 module ICU.NamesSpec
     ( spec
@@ -19,17 +19,31 @@ import qualified Unicode.Char as U
 import qualified Unicode.Char.General.Names as String
 import qualified ICU.Char as ICU
 import qualified ICU.Names as ICUString
+#ifdef HAS_BYTESTRING
+import qualified Data.ByteString.Char8 as B8
+import qualified Unicode.Char.General.Names.ByteString as ByteString
+#endif
 
 spec :: Spec
 spec = do
     describe' "name" do
         it "String" do
             traverse_ (check String.name ICUString.name) [minBound..maxBound]
+#ifdef HAS_BYTESTRING
+        it "ByteString" do
+            traverse_ (check ByteString.name (fmap B8.pack . ICUString.name)) [minBound..maxBound]
+#endif
     describe' "correctedName" do
         it "String" do
             traverse_
                 (check String.correctedName ICUString.correctedName)
                 [minBound..maxBound]
+#ifdef HAS_BYTESTRING
+        it "ByteString" do
+            traverse_
+                (check ByteString.correctedName (fmap B8.pack . ICUString.correctedName))
+                [minBound..maxBound]
+#endif
     where
     describe' = if U.unicodeVersion == ICU.unicodeVersion
         then describe