+++ unicode-data-names/test/Main.hs
@@ -4,6 +4,9 @@ module Main where
 
 import Test.Hspec ( Spec, hspec, describe )
 import qualified Unicode.Char.General.NamesSpec as String
+#ifdef HAS_BYTESTRING
+import qualified Unicode.Char.General.Names.ByteStringSpec as ByteString
+#endif
 #ifdef HAS_ICU
 import qualified ICU.NamesSpec as ICU
 #endif
@@ -14,6 +17,9 @@ main = hspec spec
 spec :: Spec
 spec = do
     describe "Unicode.Char.General.Names" String.spec
+#ifdef HAS_BYTESTRING
+    describe "Unicode.Char.General.Names.ByteString" ByteString.spec
+#endif
 #ifdef HAS_ICU
     describe "ICU.Names" ICU.spec
 #endif